+++ packages/test-app-frontend/src/App/ITwinJsApp/AppUi/PropertyComparisonFrontstage.scss
